requirements.txt
pillow==10.2.0
image==1.5.33
numpy==1.26.3
matplotlib==3.8.2